chore: move parsing unit tests
